Moved loader tests next to code
- had to edit monkeypatch logic in AWS tests due to relative
  import issues
